update: 更新文件 multicast_membership_context.cpp

+++ frameworks/js/napi/socket/async_context/src/multicast_membership_context.cpp
@@ -58,7 +58,7 @@ int MulticastMembershipContext::GetSocketFd() const
     if (sharedManager_ == nullptr) {
         return -1;
     }
-    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<intptr_t>(sharedManager_->GetData())) : -1;
+    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<uint64_t>(sharedManager_->GetData())) : -1;
 }
 
 bool MulticastMembershipContext::CheckParamsType(napi_value *params, size_t paramsCount)
